requirements.txt
pydicom==1.2.0
numpy==1.16.2
Pillow==5.3.0